Fix issue that occurs when page updates state to self
Prior to this change, when the page updated its history state to
itself (occurs when changing dark/light theme), the app treated it as
a page change (load data).
Since the page did not actually change, the app has to reload the
same data.

This change checks if the page has actually changed on a history
state update event.