Handle api request edge case
Prior to this change, a valid token that is just about to expire
could expire by the time the server handled the request. This would
give the user an error and they would have to reload the page to fix
the error.

This change handles the error by sending the request again with a
newly fetched token.