Clean up mounting of app
Prior to this change, the app was not mounted if the page was a
non-video page. We needed to listen for page change events in the
content_script in order to mount the app on a video page.

This change makes it so that we mount the app even on non-video pages.
This cleans up the app mounting process and allows for a better UX.